+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/Verifier.kt
@@ -96,6 +96,7 @@ interface Verifier {
             val issuerSigned: IssuerSigned,
         ) : VerifyCredentialResult()
 
+        @Deprecated("Has been replaced with ValidationError")
         data class InvalidStructure(
             val input: String,
             val reason: String,